+++ src/lib.rs
@@ -1,6 +1,4 @@
 pub mod commands;
 pub mod templates;
 pub mod auth;
-
-mod utils;
-pub use utils::*;
+pub mod utils;